Move `IoSlice(Mut)` tests from `std` to `coretests`
