+++ library/coretests/tests/io/mod.rs
@@ -1,1 +1,2 @@
 mod borrowed_buf;
+mod io_slice;
